Removed old DAC apply code from the detector.
